backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
firebase-admin==6.2.0
pymongo==4.6.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
scikit-image==0.22.0
scipy==1.11.4
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
bcrypt==4.0.1
